requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pypdf==5.9.0
pydantic-settings==2.10.1
redis==6.4.0
httpx==0.28.1
pytest==8.4.1
